web/requirements.txt
flask==3.0.3
requests==2.31.0
beautifulsoup4==4.9.3
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
SQLAlchemy==2.0.30
scikit-learn==1.4.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
seaborn==0.13.2
notebook==7.1.3
statsmodels==0.14.2
streamlit==1.35.0
openai==1.30.3
python-dotenv==1.0.1
plotly==5.22.0